Control.UAV.Ardupilot: Using explicit cast to avoid warning.

+++ src/Control/UAV/Ardupilot/Task.cpp
@@ -1015,7 +1015,7 @@ namespace Control
           m_fix.height = (double)gps_raw.alt * 0.001;
           m_fix.satellites = gps_raw.satellites_visible;
 
-          inf("%llu", gps_raw.time_usec);
+          inf("%llu", (long long unsigned int)gps_raw.time_usec);
 
           long time_fix = gps_raw.time_usec % 1000000000;
           unsigned int date = (unsigned int)(gps_raw.time_usec / 1e9);
